// src/app/interfaces/sysrecord.ts
export interface Chivamember {
    id: number;
    chivaCode: string;
    prefixName: string;
    firstName: string;
    lastName: string;
    createBy: string;
    bloodGroup: any;
    beAllergic: any;
    foodAllergic: any;
    congenitalDisease: any;
    moreInformation?: any;
    createDate: Date;
    updateBy: string;
    updateDate: Date;
    inActivated: boolean;
}
// ...
export interface Gift {
    id: number;
    giftNo: string;
    giftName: string;
    coverImageLocation: string;
    coverImageUrl: string;
    detailImageLocation: string;
    detailImageUrl: string;
    giftDetail: string;
    giftStartDate: Date;
    giftEndDate: Date;
    hospitalList: string;
    claimStatus: string;
    claimAmount: number;
    inActivated: boolean;
    giftApproveStatus: string;
    createBy: string;
    createDate: Date;
    updateBy: string;
    updateDate: Date;
    updateLocation: string;
}
// ...
export interface SysEnum {
    id: number;
    enumOrder: number;
    enumName: string;
    enumDesc: string;
    enumValue: string;
}



export interface SystemLog {
    id: number;
    logDateTime: Date;
    logType: number;
    errorMessage: string;
    targetSite: string;
}

export interface Giftimport {
    id: number;
    processDate: Date;
    giftId: number;
    gift: Gift;
    projectId: number;
    project: SysEnum;
    uploadQty: number;
    chivaMembers: Chivamember[];
    createBy: string;
    createDate: Date;
    updateBy: string;
    updateDate: Date;
    updateLocation: string;
}